Now notifies both a stream and pub/sub channel.
Unfortunately, blocking XREADs of Redis streams is not yet supported in
StackExchange.Redis or ServiceStack.Redis so to achieve a true "PUSH"
style processor PUB/SUB was needed.  In this revision the module pushes
a message first to the STREAM (captures the message id from the stream)
and then pushes a PUB/SUB message.  The idea here is to use the STREAM
to "catch up" in the event a client is disconnected from pub/sub and
misses messages.  At that point the pub/sub can be re-established and
continue handling published messages vs. a polling approach using
STREAMS only.